+++ crate/loading/src/system_data/sequence_component_loading_resources.rs
@@ -23,7 +23,8 @@ use map_model::loaded::{AssetMapBounds, AssetMargins};
 use sequence_model::loaded::{AssetSequenceEndTransitions, AssetWaitSequenceHandles, WaitSequence};
 use spawn_model::loaded::{AssetSpawnsSequenceHandles, Spawns, SpawnsSequence};
 use sprite_model::loaded::{
-    AssetSpritePositions, AssetSpriteRenderSequenceHandles, SpriteRenderSequence,
+    AssetScaleSequenceHandles, AssetSpritePositions, AssetSpriteRenderSequenceHandles,
+    AssetTintSequenceHandles, ScaleSequence, SpriteRenderSequence, TintSequence,
 };
 use ui_menu_item_model::loaded::AssetUiMenuItems;
 
@@ -82,6 +83,13 @@ pub struct SequenceComponentLoadingResources<'s> {
     #[derivative(Debug = "ignore")]
     pub character_cts_assets: Read<'s, AssetStorage<CharacterCts>>,
 
+    /// `TintSequence` assets.
+    #[derivative(Debug = "ignore")]
+    pub tint_sequence_assets: Read<'s, AssetStorage<TintSequence>>,
+    /// `ScaleSequence` assets.
+    #[derivative(Debug = "ignore")]
+    pub scale_sequence_assets: Read<'s, AssetStorage<ScaleSequence>>,
+
     /// `AssetSequenceEndTransitions` resource.
     #[derivative(Debug = "ignore")]
     pub asset_sequence_end_transitions: Write<'s, AssetSequenceEndTransitions>,
@@ -118,6 +126,12 @@ pub struct SequenceComponentLoadingResources<'s> {
     /// `AssetSpritePositions` resource.
     #[derivative(Debug = "ignore")]
     pub asset_sprite_positions: Write<'s, AssetSpritePositions>,
+    /// `AssetTintSequenceHandles` resource.
+    #[derivative(Debug = "ignore")]
+    pub asset_tint_sequence_handles: Write<'s, AssetTintSequenceHandles>,
+    /// `AssetScaleSequenceHandles` resource.
+    #[derivative(Debug = "ignore")]
+    pub asset_scale_sequence_handles: Write<'s, AssetScaleSequenceHandles>,
 
     /// `AssetMapBounds` resource.
     #[derivative(Debug = "ignore")]
@@ -177,6 +191,13 @@ pub struct SequenceComponentLoadingResourcesRead<'s> {
     #[derivative(Debug = "ignore")]
     pub character_cts_assets: Read<'s, AssetStorage<CharacterCts>>,
 
+    /// `TintSequence` assets.
+    #[derivative(Debug = "ignore")]
+    pub tint_sequence_assets: Read<'s, AssetStorage<TintSequence>>,
+    /// `ScaleSequence` assets.
+    #[derivative(Debug = "ignore")]
+    pub scale_sequence_assets: Read<'s, AssetStorage<ScaleSequence>>,
+
     /// `AssetSequenceEndTransitions` resource.
     #[derivative(Debug = "ignore")]
     pub asset_sequence_end_transitions: Read<'s, AssetSequenceEndTransitions>,
@@ -213,6 +234,12 @@ pub struct SequenceComponentLoadingResourcesRead<'s> {
     /// `AssetSpritePositions` resource.
     #[derivative(Debug = "ignore")]
     pub asset_sprite_positions: Read<'s, AssetSpritePositions>,
+    /// `AssetTintSequenceHandles` resource.
+    #[derivative(Debug = "ignore")]
+    pub asset_tint_sequence_handles: Read<'s, AssetTintSequenceHandles>,
+    /// `AssetScaleSequenceHandles` resource.
+    #[derivative(Debug = "ignore")]
+    pub asset_scale_sequence_handles: Read<'s, AssetScaleSequenceHandles>,
 
     /// `AssetMapBounds` resource.
     #[derivative(Debug = "ignore")]