+++ crate/sprite_model/src/loaded/asset_tint_sequence_handles.rs
@@ -1,0 +1,7 @@
+use asset_model::loaded::AssetId;
+use slotmap::SparseSecondaryMap;
+
+use crate::loaded::TintSequenceHandles;
+
+/// `TintSequenceHandles` for an asset.
+pub type AssetTintSequenceHandles = SparseSecondaryMap<AssetId, TintSequenceHandles>;